// apps/api/src/generated/enums.ts
export const ApprovalType = {
  VENDOR_CHANGE: 'VENDOR_CHANGE',
  PO_APPROVAL: 'PO_APPROVAL',
  TRANSFER_APPROVAL: 'TRANSFER_APPROVAL',
  PRICE_ADJUSTMENT: 'PRICE_ADJUSTMENT',
  STOCK_ADJUSTMENT: 'STOCK_ADJUSTMENT'
} as const

export type ApprovalType = (typeof ApprovalType)[keyof typeof ApprovalType]


export const ApprovalStatus = {
  PENDING: 'PENDING',
  APPROVED: 'APPROVED',
  REJECTED: 'REJECTED'
} as const

export type ApprovalStatus = (typeof ApprovalStatus)[keyof typeof ApprovalStatus]